Rebase PR, and allow GCR notebook image
The notebook image had been taken from the prod
leonardo docker repository. The recent changes
to notebook kernels had caused breaking changes
in the startup script. By pinning the notebook
image to the built images from this deploy
script, we now ensure that the launch script
is compatible with the target images.