Handle errors when cleaning up deleted apps/disks
If a resource was marked as deleted in Leo, but there is
an orphaned Sam record this will delete the Sam record. If
there is no Sam record (because the resource was correctly
deleted the first time) there will be a Sam permissions error.

So if a resource is marked as deleted, just log the error and continue
deleting/cleaning up the remaining resources.